requirements.txt
aiogram
beautifulsoup4
cryptography
docutils
psycopg2-binary
environs
exceptiongroup
flask-caching
flask-markdown
frozen-flask
future
gevent
lxml
pip-chill
pymorphy2
pysocks
python-dateutil
pyvirtualdisplay
pyyaml
selenium
sqlalchemy
urllib3-secure-extra
webdriver-manager
xvfbwrapper
